+++ tests/ByteDev.Collections.UnitTests/CollectionExtensionsTests.cs
@@ -20,7 +20,7 @@ namespace ByteDev.Collections.UnitTests
             [Test]
             public void WhenIsNotEmpty_ThenThrowException()
             {
-                var sut = new List<string> {"John", "Peter"};
+                ICollection<string> sut = new List<string> {"John", "Peter"};
 
                 Assert.Throws<InvalidOperationException>(() => sut.Fill(2, "Hello"));
             }
@@ -28,7 +28,7 @@ namespace ByteDev.Collections.UnitTests
             [Test]
             public void WhenNumberToFillIsLessThanOne_ThenFillNoElements()
             {
-                var sut = new List<string>();
+                ICollection<string> sut = new List<string>();
 
                 sut.Fill(0, "Hello");
 
@@ -54,14 +54,14 @@ namespace ByteDev.Collections.UnitTests
             [Test]
             public void WhenSourceIsNull_ThenThrowException()
             {
-                Assert.Throws<ArgumentNullException>(() => CollectionExtensions.IsIndexValid(null as IList<string>, 0));
+                Assert.Throws<ArgumentNullException>(() => CollectionExtensions.IsIndexValid(null as ICollection<int>, 0));
             }
 
             [TestCase(-1)]
             [TestCase(3)]
             public void WhenIndexIsOutOfRange_ThenReturnFalse(int index)
             {
-                var sut = new List<string> { "item1", "item2", "item3" };
+                ICollection<string> sut = new List<string> { "item1", "item2", "item3" };
 
                 var result = sut.IsIndexValid(index);
 
@@ -73,11 +73,101 @@ namespace ByteDev.Collections.UnitTests
             [TestCase(2)]
             public void WhenIndexIsInRange_ThenReturnTrue(int index)
             {
-                var sut = new List<string> { "item1", "item2", "item3" };
+                ICollection<string> sut = new List<string> { "item1", "item2", "item3" };
 
                 var result = sut.IsIndexValid(index);
 
                 Assert.That(result, Is.True);
+            }
+        }
+
+        [TestFixture]
+        public class AddRange : CollectionExtensionsTests
+        {
+            private readonly IEnumerable<int> CollectionToAdd = new []{ 4, 5 };
+
+            [Test]
+            public void WhenSourceIsNull_ThenThrowException()
+            {
+                Assert.Throws<ArgumentNullException>(() => CollectionExtensions.AddRange(null, CollectionToAdd));
+            }
+
+            [Test]
+            public void WhenSourceIsEmpty_ThenAddItems()
+            {
+                ICollection<int> sut = new List<int>();
+
+                sut.AddRange(CollectionToAdd);
+
+                Assert.That(sut.Count, Is.EqualTo(2));
+                Assert.That(sut.First(), Is.EqualTo(4));
+                Assert.That(sut.Second(), Is.EqualTo(5));
+            }
+
+            [Test]
+            public void WhenCollectionIsNull_ThenThrowException()
+            {
+                ICollection<int> sut = new List<int> {1, 2, 3};
+                
+                Assert.Throws<ArgumentNullException>(() => sut.AddRange(null));
+            }
+
+            [Test]
+            public void WhenCollectionNotEmpty_ThenAddItems()
+            {
+                ICollection<int> sut = new List<int> {1, 2, 3};
+
+                sut.AddRange(CollectionToAdd);
+
+                Assert.That(sut.Count, Is.EqualTo(5));
+                Assert.That(sut.First(), Is.EqualTo(1));
+                Assert.That(sut.Second(), Is.EqualTo(2));
+                Assert.That(sut.Third(), Is.EqualTo(3));
+                Assert.That(sut.Fourth(), Is.EqualTo(4));
+                Assert.That(sut.Fifth(), Is.EqualTo(5));
+            }
+        }
+
+        [TestFixture]
+        public class RemoveRange : CollectionExtensionsTests
+        {
+            private readonly IEnumerable<int> CollectionToRemove = new []{ 2, 3};
+
+            [Test]
+            public void WhenSourceIsNull_ThenThrowException()
+            {
+                Assert.Throws<ArgumentNullException>(() => CollectionExtensions.AddRange(null, CollectionToRemove));
+            }
+
+            [Test]
+            public void WhenCollectionIsNull_ThenThrowException()
+            {
+                ICollection<int> sut = new List<int> {1, 2, 3, 4, 5};
+                
+                Assert.Throws<ArgumentNullException>(() => sut.RemoveRange(null));
+            }
+
+            [Test]
+            public void WhenSourceIsEmpty_ThenRemoveNothing()
+            {
+                ICollection<int> sut = new List<int>();
+
+                sut.RemoveRange(CollectionToRemove);
+
+                Assert.That(sut, Is.Empty);
+            }
+
+            [Test]
+            public void WhenCollectionNotEmpty_ThenRemoveItems()
+            {
+                ICollection<int> sut = new List<int> {1, 2, 3, 4, 5};
+                
+                sut.RemoveRange(CollectionToRemove);
+
+                Assert.That(sut.Count, Is.EqualTo(3));
+                Assert.That(sut.First(), Is.EqualTo(1));
+                Assert.That(sut.Second(), Is.EqualTo(4));
+                Assert.That(sut.Third(), Is.EqualTo(5));
             }
         }
 
@@ -87,13 +177,13 @@ namespace ByteDev.Collections.UnitTests
             [Test]
             public void WhenSourceIsNull_ThenThrowException()
             {
-                Assert.Throws<ArgumentNullException>(() => CollectionExtensions.RemoveWhere(null as List<int>, x => x == 1));
+                Assert.Throws<ArgumentNullException>(() => CollectionExtensions.RemoveWhere(null as ICollection<int>, x => x == 1));
             }
 
             [Test]
             public void WhenPredicateIsNull_ThenThrowException()
             {
-                var sut = new List<int> {1, 2, 3};
+                ICollection<int> sut = new List<int> {1, 2, 3};
 
                 Assert.Throws<ArgumentNullException>(() => sut.RemoveWhere(null));
             }
@@ -101,7 +191,7 @@ namespace ByteDev.Collections.UnitTests
             [Test]
             public void WhenSourceIsEmpty_ThenRemoveNothing()
             {
-                var sut = new List<int>();
+                ICollection<int> sut = new List<int>();
 
                 sut.RemoveWhere(x => x == 1);
 
@@ -111,7 +201,7 @@ namespace ByteDev.Collections.UnitTests
             [Test]
             public void WhereMatchesCondition_ThenRemoveElements()
             {
-                var sut = new List<int> {1, 2, 3, 1, 2};
+                ICollection<int> sut = new List<int> {1, 2, 3, 1, 2};
 
                 sut.RemoveWhere(x => x == 1 || x == 2);