+++ docs/RELEASE-NOTES.md
@@ -10,6 +10,8 @@ New features:
 - Added `IEnumerable<T>.MaxBy`.
 - Added `IEnumerable<T>.MinBy`.
 - Added `T[,].ToSingleDimension`.
+- Added `ICollection<T>.AddRange`.
+- Added `ICollection<T>.RemoveRange`.
 
 Bug fixes / internal changes:
 - (None)